+++ CHANGES.md
@@ -1,5 +1,11 @@
 # scalding-commons #
 
+## 0.0.5
+
+* `VersionedKeyValSource` takes a reducers argument.
+* added `writeIncrementalSource` to `RichPipeEx` and `TypedRichPipeEx`, opening these methods for use with extensions of `VersionedKeyValSource` (vs the default)
+* `VersionedKeyValSource` now works in `HadoopTest` mode.
+
 ## 0.0.4
 
 * Added com.twitter.scalding.commons.RangedArgs, for accepting pairs of arguments defining some range (dates, for example)